requirements.txt
Django==5.1.*
djangorestframework==3.15.*
djangorestframework-simplejwt==5.3.*
django-cors-headers==4.4.*
django-parler==2.3.*
django-parler-rest==2.2.*
drf-spectacular==0.27.*
Pillow==11.*
python-slugify==8.*
markdown-it-py>=3.0.0
whitenoise>=6.7
requests>=2.31
